lib/test-module/python_client/requirements.txt
rich~=13.3.1
pyyaml~=6.0

protobuf
pydantic~=1.10.4

inputimeout~=1.0.4

# internal client
git+https://github.com/bringauto/internal-client-python.git@master
